Use API v2

+++ background/updater.js
@@ -1,18 +1,22 @@
 "use strict";
 
+const currentVersion = "2.2.9";
+
 (() => {
-  const { version, volume, station } = localStorage;
+  const { volume, station, version } = localStorage;
 
-  const isCurrentStationPresent = window.stationList.filter(station => station.station === station);
+  if (currentVersion > version) {
+    const isCurrentStationExist = window.stationList.some(item => station === `${item.group}.${item.station}`);
 
-  const state = {
-    version: "2.2.5",
-    volume: volume || 30,
-    state: "paused",
-    station: isCurrentStationPresent.length ? station : "tvr.kiss_fm_ua"
-  };
+    const state = {
+      version: currentVersion,
+      volume: volume || 30,
+      state: "paused",
+      station: isCurrentStationExist ? station : "TVR.KissFM"
+    };
 
-  localStorage.clear();
+    localStorage.clear();
 
-  Object.keys(state).forEach(key => localStorage.setItem(key, state[key]));
+    Object.keys(state).forEach(key => localStorage.setItem(key, state[key]));
+  }
 })();
